integrate evadv code
