+++ commands/default_cmdsets.py
@@ -17,8 +17,9 @@ own cmdsets by inheriting from them or directly from `evennia.CmdSet`.
 from evennia import default_cmds
 from evennia.contrib.grid.xyzgrid.commands import XYZGridCmdSet
 
-from commands.ooc import CmdCharCreate
-from commands.debug import DebugCmdSet
+from .debug import DebugCmdSet
+from .game import AinneveCmdSet
+from .ooc import CmdCharCreate
 
 class CharacterCmdSet(default_cmds.CharacterCmdSet):
     """
@@ -34,11 +35,10 @@ class CharacterCmdSet(default_cmds.CharacterCmdSet):
         Populates the cmdset
         """
         super().at_cmdset_creation()
-        self.add(DebugCmdSet())
-        #
-        # any commands you add below will overload the default ones.
-        #
-        self.add(XYZGridCmdSet())
+				
+        self.add(DebugCmdSet())  # commands for debugging purposes
+        self.add(XYZGridCmdSet()) # xyzgrid pathfinding and building commands
+        self.add(AinneveCmdSet()) # our own game commands
 
 class AccountCmdSet(default_cmds.AccountCmdSet):
     """
