+++ commands/ooc/chargen.py
@@ -1,5 +1,6 @@
-from evadventure.chargen import start_chargen
 from evennia.commands.default.account import CmdCharCreate as DefaultCmdCharCreate
+
+from world.chargen import start_chargen
 
 
 class CmdCharCreate(DefaultCmdCharCreate):
