--- evadventure/README.md
@@ -1,37 +1,0 @@
-# EvAdventure
-
-Contrib by Griatch 2022
-
-
-```{warning}
-NOTE - this tutorial is WIP and NOT complete! It was put on hold to focus on
-releasing Evennia 1.0. You will still learn things from it, but don't expect
-perfection.
-```
-
-A complete example MUD using Evennia. This is the final result of what is
-implemented if you follow the Getting-Started tutorial. It's recommended
-that you follow the tutorial step by step and write your own code. But if
-you prefer you can also pick apart or use this as a starting point for your
-own game.
-
-## Features
-
-- Uses a MUD-version of the [Knave](https://rpggeek.com/rpg/50827/knave) old-school
-  fantasy ruleset by Ben Milton (classless and overall compatible with early
-  edition D&D), released under the Creative Commons Attribution (all uses,
-  including commercial are allowed
-  as long as attribution is given).
-- Character creation using an editable character sheet
-- Weapons, effects, healing and resting
-- Two alternative combat systems (turn-based and twitch based)
-- Magic (three spells)
-- NPC/mobs with simple AI.
-- Simple Quest system.
-- Small game world.
-- Coded using best Evennia practices, with unit tests.
-
-
-## Installation
-
-TODO
