--- evadventure/__init__.py
@@ -1,7 +1,0 @@
-"""
-EvAdventure - a complete game in Evennia.
-
-This is an implementation of, and reference code to, the game created in the
-documentation's beginner tutorial.
-
-"""
