+++ world/shops.py
